Replace full antlr dependency with runtime
If we're using the `antlr4-maven-plugin`, then the only dependency that
should be needed is `antlr4-runtime`. I had to keep both:

* `antlr4-runtime` and
* `antlr-runtime`

because some of the source files in `cfml.parsing.antlr` depend on the
`3.5.2` runtime.

A few test classes depend on classes in the main `antlr4` library, so I
added that one to the `test` scope, so that it doesn't get included in
the final package.